requirements.txt
flask
requests
beautifulsoup4
lxml
pytest
firecrawl-py
python-dotenv
